requirements.txt
Django==3.1.7
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-crispy-forms==1.11.1
django-js-asset==1.2.2
